Add separated files for screens
