+++ include/screens/disclaimer.h
@@ -1,0 +1,9 @@
+#ifndef SCREENS_DISCLAIMER_H
+#define SCREENS_DISCLAIMER_H
+
+void screen_disclaimer_load();
+void screen_disclaimer_unload();
+void screen_disclaimer_update();
+void screen_disclaimer_draw();
+
+#endif
