+++ include/screens/levelselect.h
@@ -1,0 +1,9 @@
+#ifndef SCREENS_LEVELSELECT_H
+#define SCREENS_LEVELSELECT_H
+
+void screen_levelselect_load();
+void screen_levelselect_unload();
+void screen_levelselect_update();
+void screen_levelselect_draw();
+
+#endif
