+++ src/screen_levelselect.c
@@ -1,0 +1,68 @@
+#include "screens/levelselect.h"
+#include <stdint.h>
+#include <string.h>
+#include <stdio.h>
+
+#include "input.h"
+#include "screen.h"
+#include "render.h"
+#include "screens/level.h"
+
+#define MAX_LEVELS 4
+
+static uint8_t menu_choice = 0;
+
+void screen_levelselect_load()
+{
+    menu_choice = 0;
+}
+
+void screen_levelselect_unload() {}
+
+void screen_levelselect_update()
+{
+    if(pad_pressed(PAD_DOWN))
+            menu_choice++;
+        else if(pad_pressed(PAD_UP)) {
+            if(menu_choice == 0) menu_choice = MAX_LEVELS - 1;
+            else menu_choice--;
+        }
+
+        menu_choice = (menu_choice % MAX_LEVELS);
+
+        if(pad_pressed(PAD_START) || pad_pressed(PAD_CROSS)) {
+            screen_level_setlevel(menu_choice);
+            scene_change(SCREEN_LEVEL);
+        }
+}
+
+void screen_levelselect_draw()
+{
+    char buffer[255] = { 0 };
+    
+    int16_t x;
+    const char *title = "*** engine-psx ***";
+    x = CENTERX - (strlen(title) * 4);
+    draw_text(x, 12, 0, title);
+
+    const char *subtitle = "https://luksamuk.codes/";
+    x = CENTERX - (strlen(subtitle) * 4);
+    draw_text(x, 24, 0, subtitle);
+
+    snprintf(buffer, 255, "%s %s", __DATE__, __TIME__);
+    x = SCREEN_XRES - (strlen(buffer) * 8) - 8;
+    draw_text(x, SCREEN_YRES - 24, 0, buffer);
+
+    snprintf(
+        buffer, 255,
+        "%c Round 0 Zone 1\n"
+        "%c Round 0 Zone 2\n"
+        "%c Round 1 Zone 1\n"
+        "%c Round 1 Zone 2\n",
+        (menu_choice == 0) ? '>' : ' ',
+        (menu_choice == 1) ? '>' : ' ',
+        (menu_choice == 2) ? '>' : ' ',
+        (menu_choice == 3) ? '>' : ' ');
+    draw_text(8, 60, 0, buffer);
+}
+
